requirements-dev.txt
click
Pillow